+++ ReactComponents/src/styles/NavbarStyles.ts
