+++ ReactComponents/src/views/TicTacToe/types/moves.ts
@@ -1,0 +1,10 @@
+export enum Moves {
+  X = "X",
+  O = "O",
+}
+
+export enum Draw {
+  D = "D",
+}
+
+export type Result = Moves | Draw;
